Time: 11 ms (35.35%), Space: 23 MB (45.85%) - LeetHub

+++ 0875-koko-eating-bananas/0875-koko-eating-bananas.cpp
@@ -1,18 +1,18 @@
 class Solution {
 public:
-    bool helper(vector<int>& piles,int mid,int h){
+    bool helper(vector<int>&piles,long long mid,int h){
         long long hours=0;
         for(int i=0; i<piles.size(); i++){
-            hours+=(long)(piles[i]+mid-1)/(long)mid;
+            hours+=(long)(piles[i]+mid-1)/mid;
         }
         return hours<=h;
     }
     int minEatingSpeed(vector<int>& piles, int h) {
-        long left=1;
-        long right=1e9;
-        long ans=0;
+        long long left=1;
+        long long right=1e9;
+        long long ans=0;
         while(left<=right){
-            long mid=left+(right-left)/2;
+            long long mid=left+(right-left)/2;
             if(helper(piles,mid,h)){
                 ans=mid;
                 right=mid-1;
